pyproject.toml
[tool.poetry]
name = "changelogged"
version = "0.7.0"
description = "Automated management of your changelog and other versioned files, following the principles of Keep a Changelog and Semantic Versioning."
license = "MIT"
authors = ["award28 <austin.ward@klaviyo.com>"]
maintainers = ["award28 <austin.ward@klaviyo.com>"]
readme = "docs/README.md"
repository = "https://github.com/award28/changelogger"
packages = [
    { include = "changelogger" },
]
include = [
    { path = "assets" },
]
keywords = ["changelog", "version", "semver", "changelogger"]

[tool.poetry.scripts]
changelogger = "changelogger.__main__:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = "0.7.0"
rich = "12.6.0"
pydantic = "1.10.2"
semver = "2.13.0"
jinja2 = "3.1.2"
pyyaml = "6.0"
gitpython = "3.1.30"
pytest = "^7.2.1"
freezegun = "^1.2.2"
mypy = "^0.991"
coverage = "^7.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.pycln]
all = true

[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.coverage]
    [tool.coverage.run]
        source = ["changelogger"]
